Don't require relocs at run time
Our metavector used to refer to reloc records that had been retained
(with -Wl,-q). This is poor value in terms of binary size. Now instead
we use a different metavector record format for reloc targets, which
keeps the base address and size info in the record itself (there is no
type, by the nature of reloc targets), and forget about mapping the
relocs.